✨ feat(kernel): 建立设备运行时、抢占调度与领域错误体系
* 建立固件设备目录与硬件资源模型
    - 新增启动期只读 Catalog，以固定容量快照统一保存固件设备、CPU、interrupt controller 与平台内存图和时基事实
    - 实现 FDT 枚举后端，解析 CPU、compatible、reg、status、interrupt-controller 与 timebase-frequency，并以稳定 FirmwareId、父子关系和资源重叠校验发布目录
    - 保留 ACPI Enumerator 边界，引入设备分类、绑定策略和 controller role，使后续内核驱动与用户态驱动能够共享同一发现模型
    - 新增 MMIO capability 对象，通过 KernelMM 事务化映射和解除映射页对齐物理区间，并将 MMIO 对象类型纳入公共 capability ABI

* 引入 RCU 驱动的通用中断框架与 PLIC
    - 为 Tay 新增 policy-based rcu_domain、read guard、固定容量 retire 批次和同步 grace period，在硬中断读侧保持无分配、无阻塞访问
    - 以 copy-on-write Registry 和 RCU 快照管理 trap source、IRQ handler、domain 与 cascade 绑定，使用 generation token 拒绝陈旧取消并在注销后保证 context 可立即销毁
    - 建立与固件控制器无关的 IRQ domain flow，统一 resolve、claim、ack、EOI、complete、mask、unmask 和 priority，并限制级联深度与单次 trap claim 数量
    - 实现 RISC-V PLIC 驱动，从 Catalog 获取控制器与 MMIO、选择 BSP context、注册 root domain，并将双架构异步 trap 接入公共 dispatcher

* 重构 BSP Scheduler 为事件驱动的 RR/FIFO 核心
    - 引入 per-CPU RunQueue、内嵌 SchedulerStorage、Thread owner Adapter、独立 accounting 及无虚调用的 enter、leave、select 调度类
    - 让 RunQueue::current 成为 current 的唯一来源，ready queue 只保存 SchedEntity，并在提交切换时统一恢复 Thread、地址空间与延迟退出所有权
    - 将 attach、wake、block、yield、preempt 和 exit 收敛为 SchedulerCore 生命周期事件，所有可恢复 trap 只在统一返回边界执行 schedule
    - 增加 BLOCKING 状态、wake_pending 与线性 BlockToken，以 prepare、commit、cancel 两阶段 park 协议消除 WorkQueue 和 timed wait 的丢失唤醒窗口

* 引入 precision timer、异步 WorkQueue 与 RR 抢占
    - 为 Tay 新增无分配 intrusive pairing heap，并扩展 units::time/duration、比较、算术、字面量和 saturated_add，为绝对 deadline 与运行时间记账提供强类型基础
    - 新增 DeadlineState 合并 precision timer root 与 Scheduler RR 来源，在 RISC-V64 SBI 和 LoongArch64 CSR 后端始终编程最早 one-shot deadline
    - 实现 PrecisionTimerEngine、宿主嵌入 TimerNode 和 typed Worklet；timer IRQ 只摘取到期节点并 post completion，不在中断中执行虚函数或销毁宿主
    - 实现单 worker WorkQueue 的 IRQ-safe post、固定预算 drain、两阶段 park、self-repost、tail-dispatch、统计与 shutdown，并启动永久 BSP queue
    - 为 RR 线程发布 10 ms 抢占 deadline，使 CPU-bound kernel Thread 可被硬件 timer 抢占；无竞争者时自动停用该来源
    - 为 Thread 增加 generation 与 pin 保护的 wait_until，统一处理 timeout、wake、cancel、重复等待和竞态清理，并以周期 Worklet 验证重新 arm 路径

* 建立跨子系统结构化领域错误
    - 新增 2-byte trivially-copyable KernelError，以 Type(Reason) 稳定编码跨领域错误，并提供 reason 查询、Tay error 映射和统一 formatter
    - 为 capability、paging、memory segment、kernel layout、Process、Thread、Scheduler、timer、WorkQueue、usrboot、Catalog、FDT、MMIO 与 PLIC 定义携带上下文 payload 的 variant error
    - 将内存管理、对象模型、调度、设备发现、启动和异步 API 迁移为 typed expected，在抽象边界保留嵌套 cause、具体失败对象与生命周期状态
    - 扩展 Tay expected 的 TAY_TRY、TAY_TRYV 和 TAY_ERR 单次求值传播接口，并修正引用值提取、variant visit 返回类型与 formatter 组合能力

* 完善 Tay 无异常泛型基础设施
    - 新增 freestanding tay::optional、nullopt、in_place、value_or、emplace 与契约 panic，供 KernelError reason 查询和可选状态表达复用
    - 抽取状态化 projected_compare 并复用于 introsort 与 timer heap，保持 Compare/Projection 的值类别、EBO 与 noexcept 传播
    - 为 RCU、optional、expected、pairing heap、units 和 sort 增加 Host、panic、公开头与 freestanding 契约覆盖
    - 更新 Tay callable、侵入式容器和错误处理文档，明确非拥有节点、grace period、失败传播与时间算术边界

* 统一内核启动验证、测试夹具与架构边界
    - 新增按初始化阶段显式注册的 kernel selftest framework，在 timer、Scheduler、BSP WorkQueue 和 usrboot 提交边界串行运行用例
    - 覆盖领域错误编码与 payload、RCU 中断注册表、capability、两阶段 park、RR 抢占、WorkQueue 生命周期、precision timer、timed wait 和 FIFO handoff
    - 将 IntegerObject 从生产对象模型收敛为 selftest 专用夹具，统一 capability 与 IRQ 测试入口，避免生产构建携带测试对象
    - 更新内核 README、构建登记、架构 namespace 与 SBI/LABOOT section 宏命名，明确当前 BSP-only 调度、timer、WorkQueue 和设备运行边界

+++ kernel/boot/early.cpp
@@ -177,7 +177,7 @@ extern "C" [[noreturn]] BOOT_INIT_TEXT void __bsp_early_main(
     memory::init_kernel_space(*boot::context().info);
     auto kernel_mm = memory::KernelMM::initialize(*boot::context().info);
     if (!kernel_mm)
-        kernel::log::panic("无法初始化 KernelMM: {}", static_cast<int>(kernel_mm.error()));
+        kernel::log::panic("无法初始化 KernelMM: {}", kernel_mm.error());
     for (size_t index = 0; index < memory::page_database().region_count(); ++index) {
         const auto &area = memory::page_database().region(index).parent;
         if (!memory::kernel_mm().hhdm_covers(area.begin, area.size()))
